+++ packages/common/evmUtils/src/operations/transaction/index.ts
@@ -1,2 +1,3 @@
 export * from './getTransactionOperation';
 export * from './getWalletTransactionsOperation';
+export * from './getWalletTransactionsVerboseOperation';