+++ packages/client/firebaseSolAuth/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @moralisweb3/client-firebase-sol-auth
 
+## 2.8.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/client-firebase-auth-utils@2.8.2
+
+## 2.8.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/client-firebase-auth-utils@2.8.1
+
 ## 2.8.0
 
 ### Minor Changes