+++ packages/evmApi/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @moralisweb3/evm-api
 
+## 2.8.2
+
+### Patch Changes
+
+- Updated dependencies [[`0cefcef6`](https://github.com/MoralisWeb3/Moralis-JS-SDK/commit/0cefcef6f42cd4f37e2c7ca0d0499761750e114b), [`0cefcef6`](https://github.com/MoralisWeb3/Moralis-JS-SDK/commit/0cefcef6f42cd4f37e2c7ca0d0499761750e114b)]:
+  - @moralisweb3/common-evm-utils@2.8.2
+  - @moralisweb3/api-utils@2.8.2
+  - @moralisweb3/common-core@2.8.2
+
+## 2.8.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/api-utils@2.8.1
+  - @moralisweb3/common-core@2.8.1
+  - @moralisweb3/common-evm-utils@2.8.1
+
 ## 2.8.0
 
 ### Minor Changes